:recycle: [Refactor] Developer -> Auth 로 회원탈퇴 로직 이동
