+++ src/main/java/ssuPlector/controller/AuthController.java
@@ -2,6 +2,7 @@ package ssuPlector.controller;
 
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -18,6 +19,7 @@ import ssuPlector.global.response.ApiResponse;
 import ssuPlector.security.handler.annotation.AuthUser;
 import ssuPlector.security.handler.annotation.ExtractToken;
 import ssuPlector.service.auth.AuthService;
+import ssuPlector.validation.annotation.ExistDeveloper;
 
 @RestController
 @RequiredArgsConstructor
@@ -44,5 +46,13 @@ public class AuthController {
     @PostMapping("/kakao/refresh")
     public ApiResponse<TokenRefreshResponse> refresh(@ExtractToken String refreshToken) {
         return ApiResponse.onSuccess("토큰 재발급 완료", authService.refresh(refreshToken));
+    }
+
+    @Operation(summary = "회원 탈퇴", description = "회원을 비활성화 합니다(soft delete)._현근")
+    @DeleteMapping("/{developerId}")
+    public ApiResponse<String> withdrawDeveloper(
+            @ExistDeveloper @PathVariable("developerId") Long developerId) {
+        authService.withdrawDeveloper(developerId);
+        return ApiResponse.onSuccess("회원 탈퇴 성공");
     }
 }
