+++ src/main/java/ssuPlector/service/developer/DeveloperService.java
@@ -20,6 +20,4 @@ public interface DeveloperService {
     List<Long> getUpdateTargetDeveloperIds(List<Long> DeveloperIdList);
 
     Page<Developer> getDeveloperList(DeveloperListRequestDTO requestDTO, int page);
-
-    void withdrawDeveloper(Long developerId);
 }
